+++ PACECloudSDK/POIKit/GeoService/GeoAPIManagerError.swift
@@ -1,0 +1,15 @@
+//
+//  GeoAPIManagerError.swift
+//  PACECloudSDK
+//
+//  Created by PACE Telematics GmbH.
+//
+
+import Foundation
+
+enum GeoApiManagerError: Error {
+    case invalidSpeed
+    case requestCancelled
+    case invalidResponse
+    case unknownError
+}