+++ PACECloudSlimSDK/GeoAPIManager+Slim.swift
@@ -14,11 +14,11 @@ class GeoAPIManager {
 
     func fetchPolygons() {}
 
-    func locationBasedCofuStations(for location: CLLocation, result: @escaping (Result<[AppKit.CofuGasStation], GeoApiManagerError>) -> Void) {
+    func locationBasedCofuStations(for location: CLLocation, result: @escaping (Result<[POIKit.CofuGasStation], GeoApiManagerError>) -> Void) {
         result(.failure(.unknownError))
     }
 
-    func cofuGasStations(option: AppKit.CofuGasStation.Option, result: @escaping (Result<[AppKit.CofuGasStation], GeoApiManagerError>) -> Void) {
+    func cofuGasStations(option: POIKit.CofuGasStation.Option, result: @escaping (Result<[POIKit.CofuGasStation], GeoApiManagerError>) -> Void) {
         result(.failure(.unknownError))
     }