+++ PACECloudSlimSDK/POIKitAPIProtocol+Slim.swift
@@ -9,26 +9,24 @@ import CoreLocation
 import Foundation
 
 extension POIKitAPIProtocol {
-    func fetchPOIs(poisOfType: POIKit.POILayer,
-                   boundingBox: POIKit.BoundingBox,
+    func fetchPOIs(boundingBox: POIKit.BoundingBox,
                    forceLoad: Bool = false,
                    handler: @escaping (Result<[POIKit.GasStation], Error>) -> Void) -> CancellablePOIAPIRequest? {
         nil
     }
 
-    func loadPOIs(poisOfType: POIKit.POILayer,
-                  boundingBox: POIKit.BoundingBox,
+    func fetchPOIs(locations: [CLLocation],
+                  handler: @escaping (Result<[POIKit.GasStation], Error>) -> Void) -> CancellablePOIAPIRequest? {
+        nil
+    }
+
+    func loadPOIs(boundingBox: POIKit.BoundingBox,
                   forceLoad: Bool = false,
                   handler: @escaping (Result<[POIKit.GasStation], Error>) -> Void) -> CancellablePOIAPIRequest? {
         nil
     }
 
     func loadPOIs(uuids: [String],
-                  handler: @escaping (Result<[POIKit.GasStation], Error>) -> Void) -> CancellablePOIAPIRequest? {
-        nil
-    }
-
-    func loadPOIs(locations: [CLLocation],
                   handler: @escaping (Result<[POIKit.GasStation], Error>) -> Void) -> CancellablePOIAPIRequest? {
         nil
     }